+++ example_message_interface.py
@@ -11,11 +11,10 @@ host|host:port            (optional) Set the address (or using a colon,
                           the address and port). Defaults to a hard-coded test
                           address and port.
 '''
-from openlcb.tcpsocket import TcpSocket
+from canbus.tcpsocket import TcpSocket
 
 from canbus.canphysicallayergridconnect import CanPhysicalLayerGridConnect
 from canbus.canlink import CanLink
-# from openlcb.controlframe import ControlFrame
 from openlcb.nodeid import NodeID
 from openlcb.message import Message
 from openlcb.mti import MTI
